--- jit/corePlugins/backgroundBlendMode.js
@@ -1,1 +1,0 @@
-module.exports = require('../../lib/plugins/backgroundBlendMode').default()
