--- jit/corePlugins/fontSmoothing.js
@@ -1,12 +1,0 @@
-const { createSimpleStaticUtilityPlugin } = require('../pluginUtils')
-
-module.exports = createSimpleStaticUtilityPlugin({
-  '.antialiased': {
-    '-webkit-font-smoothing': 'antialiased',
-    '-moz-osx-font-smoothing': 'grayscale',
-  },
-  '.subpixel-antialiased': {
-    '-webkit-font-smoothing': 'auto',
-    '-moz-osx-font-smoothing': 'auto',
-  },
-})
