+++ jit/corePlugins/index.js
@@ -286,24 +286,24 @@ module.exports = {
   fontStyle: require('../../lib/plugins/fontStyle').default(),
   fontVariantNumeric: require('../../lib/plugins/fontVariantNumeric').default(),
   lineHeight: require('../../lib/plugins/lineHeight').default(),
-  letterSpacing: require('./letterSpacing'),
-  textColor: require('./textColor'),
-  textOpacity: require('./textOpacity'),
-  textDecoration: require('./textDecoration'),
-  fontSmoothing: require('./fontSmoothing'),
-  placeholderColor: require('./placeholderColor'),
-  placeholderOpacity: require('./placeholderOpacity'),
+  letterSpacing: require('../../lib/plugins/letterSpacing').default(),
+  textColor: require('../../lib/plugins/textColor').default(),
+  textOpacity: require('../../lib/plugins/textOpacity').default(),
+  textDecoration: require('../../lib/plugins/textDecoration').default(),
+  fontSmoothing: require('../../lib/plugins/fontSmoothing').default(),
+  placeholderColor: require('../../lib/plugins/placeholderColor').default(),
+  placeholderOpacity: require('../../lib/plugins/placeholderOpacity').default(),
 
-  opacity: require('./opacity'),
-  backgroundBlendMode: require('./backgroundBlendMode'),
-  mixBlendMode: require('./mixBlendMode'),
-  boxShadow: require('./boxShadow'),
-  outline: require('./outline'),
-  ringWidth: require('./ringWidth'),
-  ringColor: require('./ringColor'),
-  ringOpacity: require('./ringOpacity'),
-  ringOffsetWidth: require('./ringOffsetWidth'),
-  ringOffsetColor: require('./ringOffsetColor'),
+  opacity: require('../../lib/plugins/opacity').default(),
+  backgroundBlendMode: require('../../lib/plugins/backgroundBlendMode').default(),
+  mixBlendMode: require('../../lib/plugins/mixBlendMode').default(),
+  boxShadow: require('../../lib/plugins/boxShadow').default(),
+  outline: require('../../lib/plugins/outline').default(),
+  ringWidth: require('../../lib/plugins/ringWidth').default(),
+  ringColor: require('../../lib/plugins/ringColor').default(),
+  ringOpacity: require('../../lib/plugins/ringOpacity').default(),
+  ringOffsetWidth: require('../../lib/plugins/ringOffsetWidth').default(),
+  ringOffsetColor: require('../../lib/plugins/ringOffsetColor').default(),
 
   filter: require('../../lib/plugins/filter').default(),
   blur: require('../../lib/plugins/blur').default(),
@@ -327,8 +327,8 @@ module.exports = {
   backdropSaturate: require('../../lib/plugins/backdropSaturate').default(),
   backdropSepia: require('../../lib/plugins/backdropSepia').default(),
 
-  transitionProperty: require('./transitionProperty'),
-  transitionDelay: require('./transitionDelay'),
-  transitionDuration: require('./transitionDuration'),
-  transitionTimingFunction: require('./transitionTimingFunction'),
+  transitionProperty: require('../../lib/plugins/transitionProperty').default(),
+  transitionDelay: require('../../lib/plugins/transitionDelay').default(),
+  transitionDuration: require('../../lib/plugins/transitionDuration').default(),
+  transitionTimingFunction: require('../../lib/plugins/transitionTimingFunction').default(),
 }
