--- jit/corePlugins/mixBlendMode.js
@@ -1,1 +1,0 @@
-module.exports = require('../../lib/plugins/mixBlendMode').default()
