--- jit/corePlugins/transitionDelay.js
@@ -1,17 +1,0 @@
-const { nameClass } = require('../pluginUtils')
-
-module.exports = function ({ matchUtilities }) {
-  matchUtilities({
-    delay: (modifier, { theme }) => {
-      let value = theme.transitionDelay[modifier]
-
-      if (value === undefined) {
-        return []
-      }
-
-      return {
-        [nameClass('delay', modifier)]: { 'transition-delay': value },
-      }
-    },
-  })
-}
