+++ jit/tests/arbitrary-values.test.html
@@ -93,5 +93,17 @@
     <div class="fill-[var(--color)]"></div>
     <div class="stroke-[#da5b66]"></div>
     <div class="leading-[var(--leading)]"></div>
+    <div class="tracking-[var(--tracking)]"></div>
+    <div class="placeholder-[var(--placeholder)]"></div>
+    <div class="placeholder-opacity-[var(--placeholder-opacity)]"></div>
+    <div class="opacity-[var(--opacity)]"></div>
+    <div class="outline-[var(--outline)]"></div>
+    <div class="ring-[#76ad65]"></div>
+    <div class="ring-offset-[#76ad65]"></div>
+    <div class="ring-[10px]"></div>
+    <div class="ring-offset-[#ad672f]"></div>
+    <div class="ring-offset-[19rem]"></div>
+    <div class="ring-opacity-[var(--ring-opacity)]"></div>
+    <div class="delay-[var(--delay)]"></div>
   </body>
 </html>
