+++ jit/tests/basic-usage.test.html
@@ -31,6 +31,7 @@
     <div class="border-solid"></div>
     <div class="border"></div>
     <div class="border-2"></div>
+    <div class="shadow"></div>
     <div class="shadow-md"></div>
     <div class="shadow-lg"></div>
     <div class="decoration-clone decoration-slice"></div>
