+++ src/plugins/boxShadow.js
@@ -1,33 +1,67 @@
 import _ from 'lodash'
-import nameClass from '../util/nameClass'
+import transformThemeValue from '../util/transformThemeValue'
+const { nameClass } = require('../../jit/pluginUtils')
+
+let transformValue = transformThemeValue('boxShadow')
+let shadowReset = {
+  '*': {
+    '--tw-shadow': '0 0 #0000',
+  },
+}
 
 export default function () {
-  return function ({ addUtilities, theme, variants }) {
-    addUtilities(
-      {
-        '*': {
-          '--tw-shadow': '0 0 #0000',
+  return function ({ config, matchUtilities, addBase, addUtilities, theme, variants }) {
+    if (config('mode') === 'jit') {
+      addBase(shadowReset)
+      matchUtilities({
+        shadow: (modifier, { theme }) => {
+          let value = transformValue(theme.boxShadow[modifier])
+
+          if (value === undefined) {
+            return []
+          }
+
+          return [
+            {
+              [nameClass('shadow', modifier)]: {
+                '--tw-shadow': value === 'none' ? '0 0 #0000' : value,
+                'box-shadow': [
+                  `var(--tw-ring-offset-shadow, 0 0 #0000)`,
+                  `var(--tw-ring-shadow, 0 0 #0000)`,
+                  `var(--tw-shadow)`,
+                ].join(', '),
+              },
+            },
+          ]
         },
-      },
-      { respectImportant: false }
-    )
-
-    const utilities = _.fromPairs(
-      _.map(theme('boxShadow'), (value, modifier) => {
-        return [
-          nameClass('shadow', modifier),
-          {
-            '--tw-shadow': value === 'none' ? '0 0 #0000' : value,
-            'box-shadow': [
-              `var(--tw-ring-offset-shadow, 0 0 #0000)`,
-              `var(--tw-ring-shadow, 0 0 #0000)`,
-              `var(--tw-shadow)`,
-            ].join(', '),
-          },
-        ]
       })
-    )
+    } else {
+      addUtilities(
+        {
+          '*': {
+            '--tw-shadow': '0 0 #0000',
+          },
+        },
+        { respectImportant: false }
+      )
 
-    addUtilities(utilities, variants('boxShadow'))
+      const utilities = _.fromPairs(
+        _.map(theme('boxShadow'), (value, modifier) => {
+          return [
+            nameClass('shadow', modifier),
+            {
+              '--tw-shadow': value === 'none' ? '0 0 #0000' : value,
+              'box-shadow': [
+                `var(--tw-ring-offset-shadow, 0 0 #0000)`,
+                `var(--tw-ring-shadow, 0 0 #0000)`,
+                `var(--tw-shadow)`,
+              ].join(', '),
+            },
+          ]
+        })
+      )
+
+      addUtilities(utilities, variants('boxShadow'))
+    }
   }
 }
