+++ src/plugins/placeholderColor.js
@@ -1,31 +1,53 @@
 import _ from 'lodash'
 import flattenColorPalette from '../util/flattenColorPalette'
-import nameClass from '../util/nameClass'
 import toColorValue from '../util/toColorValue'
 import withAlphaVariable from '../util/withAlphaVariable'
+const { asValue, nameClass } = require('../../jit/pluginUtils')
 
 export default function () {
-  return function ({ addUtilities, theme, variants, corePlugins }) {
-    const colors = flattenColorPalette(theme('placeholderColor'))
+  return function ({ config, matchUtilities, addUtilities, theme, variants, corePlugins }) {
+    if (config('mode') === 'jit') {
+      let colorPalette = flattenColorPalette(theme('placeholderColor'))
 
-    const getProperties = (value) => {
-      if (corePlugins('placeholderOpacity')) {
-        return withAlphaVariable({
-          color: value,
-          property: 'color',
-          variable: '--tw-placeholder-opacity',
-        })
+      matchUtilities({
+        placeholder: (modifier) => {
+          let value = asValue(modifier, colorPalette)
+
+          if (value === undefined) {
+            return []
+          }
+
+          return {
+            [`${nameClass('placeholder', modifier)}::placeholder`]: withAlphaVariable({
+              color: value,
+              property: 'color',
+              variable: '--tw-placeholder-opacity',
+            }),
+          }
+        },
+      })
+    } else {
+      const colors = flattenColorPalette(theme('placeholderColor'))
+
+      const getProperties = (value) => {
+        if (corePlugins('placeholderOpacity')) {
+          return withAlphaVariable({
+            color: value,
+            property: 'color',
+            variable: '--tw-placeholder-opacity',
+          })
+        }
+
+        return { color: toColorValue(value) }
       }
 
-      return { color: toColorValue(value) }
+      const utilities = _.fromPairs(
+        _.map(colors, (value, modifier) => {
+          return [`${nameClass('placeholder', modifier)}::placeholder`, getProperties(value)]
+        })
+      )
+
+      addUtilities(utilities, variants('placeholderColor'))
     }
-
-    const utilities = _.fromPairs(
-      _.map(colors, (value, modifier) => {
-        return [`${nameClass('placeholder', modifier)}::placeholder`, getProperties(value)]
-      })
-    )
-
-    addUtilities(utilities, variants('placeholderColor'))
   }
 }
