+++ src/plugins/placeholderOpacity.js
@@ -1,19 +1,37 @@
 import _ from 'lodash'
-import nameClass from '../util/nameClass'
+const { asValue, nameClass } = require('../../jit/pluginUtils')
 
 export default function () {
-  return function ({ addUtilities, theme, variants }) {
-    const utilities = _.fromPairs(
-      _.map(theme('placeholderOpacity'), (value, modifier) => {
-        return [
-          `${nameClass('placeholder-opacity', modifier)}::placeholder`,
-          {
-            '--tw-placeholder-opacity': value,
-          },
-        ]
-      })
-    )
+  return function ({ config, matchUtilities, addUtilities, theme, variants }) {
+    if (config('mode') === 'jit') {
+      matchUtilities({
+        'placeholder-opacity': (modifier, { theme }) => {
+          let value = asValue(modifier, theme.placeholderOpacity)
 
-    addUtilities(utilities, variants('placeholderOpacity'))
+          if (value === undefined) {
+            return []
+          }
+
+          return {
+            [`${nameClass('placeholder-opacity', modifier)}::placeholder`]: {
+              '--tw-placeholder-opacity': value,
+            },
+          }
+        },
+      })
+    } else {
+      const utilities = _.fromPairs(
+        _.map(theme('placeholderOpacity'), (value, modifier) => {
+          return [
+            `${nameClass('placeholder-opacity', modifier)}::placeholder`,
+            {
+              '--tw-placeholder-opacity': value,
+            },
+          ]
+        })
+      )
+
+      addUtilities(utilities, variants('placeholderOpacity'))
+    }
   }
 }
