+++ src/plugins/ringOffsetColor.js
@@ -1,21 +1,41 @@
 import _ from 'lodash'
 import flattenColorPalette from '../util/flattenColorPalette'
-import nameClass from '../util/nameClass'
 import toColorValue from '../util/toColorValue'
+const { asColor, nameClass } = require('../../jit/pluginUtils')
 
 export default function () {
-  return function ({ addUtilities, theme, variants }) {
-    const colors = flattenColorPalette(theme('ringOffsetColor'))
-    const utilities = _.fromPairs(
-      _.map(_.omit(colors, 'DEFAULT'), (value, modifier) => {
-        return [
-          nameClass('ring-offset', modifier),
-          {
-            '--tw-ring-offset-color': toColorValue(value),
-          },
-        ]
+  return function ({ config, matchUtilities, addUtilities, theme, variants }) {
+    if (config('mode') === 'jit') {
+      let colorPalette = flattenColorPalette(theme('ringOffsetColor'))
+
+      matchUtilities({
+        'ring-offset': (modifier) => {
+          let value = asColor(modifier, colorPalette)
+
+          if (value === undefined) {
+            return []
+          }
+
+          return {
+            [nameClass('ring-offset', modifier)]: {
+              '--tw-ring-offset-color': toColorValue(value),
+            },
+          }
+        },
       })
-    )
-    addUtilities(utilities, variants('ringOffsetColor'))
+    } else {
+      const colors = flattenColorPalette(theme('ringOffsetColor'))
+      const utilities = _.fromPairs(
+        _.map(_.omit(colors, 'DEFAULT'), (value, modifier) => {
+          return [
+            nameClass('ring-offset', modifier),
+            {
+              '--tw-ring-offset-color': toColorValue(value),
+            },
+          ]
+        })
+      )
+      addUtilities(utilities, variants('ringOffsetColor'))
+    }
   }
 }
