+++ src/plugins/ringOffsetWidth.js
@@ -1,18 +1,8 @@
-import _ from 'lodash'
-import nameClass from '../util/nameClass'
+import createUtilityPlugin from '../util/createUtilityPlugin'
+const { asLength } = require('../../jit/pluginUtils')
 
 export default function () {
-  return function ({ addUtilities, theme, variants }) {
-    const utilities = _.fromPairs(
-      _.map(_.omit(theme('ringOffsetWidth'), 'DEFAULT'), (value, modifier) => {
-        return [
-          nameClass('ring-offset', modifier),
-          {
-            '--tw-ring-offset-width': value,
-          },
-        ]
-      })
-    )
-    addUtilities(utilities, variants('ringOffsetWidth'))
-  }
+  return createUtilityPlugin('ringOffsetWidth', [['ring-offset', ['--tw-ring-offset-width']]], {
+    resolveArbitraryValue: asLength,
+  })
 }
