+++ src/plugins/textColor.js
@@ -1,31 +1,53 @@
 import _ from 'lodash'
 import flattenColorPalette from '../util/flattenColorPalette'
-import nameClass from '../util/nameClass'
 import toColorValue from '../util/toColorValue'
 import withAlphaVariable from '../util/withAlphaVariable'
+const { asColor, nameClass } = require('../../jit/pluginUtils')
 
 export default function () {
-  return function ({ addUtilities, theme, variants, corePlugins }) {
-    const colors = flattenColorPalette(theme('textColor'))
+  return function ({ config, matchUtilities, addUtilities, theme, variants, corePlugins }) {
+    if (config('mode') === 'jit') {
+      let colorPalette = flattenColorPalette(theme('textColor'))
 
-    const getProperties = (value) => {
-      if (corePlugins('textOpacity')) {
-        return withAlphaVariable({
-          color: value,
-          property: 'color',
-          variable: '--tw-text-opacity',
-        })
+      matchUtilities({
+        text: (modifier) => {
+          let value = asColor(modifier, colorPalette)
+
+          if (value === undefined) {
+            return []
+          }
+
+          return {
+            [nameClass('text', modifier)]: withAlphaVariable({
+              color: value,
+              property: 'color',
+              variable: '--tw-text-opacity',
+            }),
+          }
+        },
+      })
+    } else {
+      const colors = flattenColorPalette(theme('textColor'))
+
+      const getProperties = (value) => {
+        if (corePlugins('textOpacity')) {
+          return withAlphaVariable({
+            color: value,
+            property: 'color',
+            variable: '--tw-text-opacity',
+          })
+        }
+
+        return { color: toColorValue(value) }
       }
 
-      return { color: toColorValue(value) }
+      const utilities = _.fromPairs(
+        _.map(colors, (value, modifier) => {
+          return [nameClass('text', modifier), getProperties(value)]
+        })
+      )
+
+      addUtilities(utilities, variants('textColor'))
     }
-
-    const utilities = _.fromPairs(
-      _.map(colors, (value, modifier) => {
-        return [nameClass('text', modifier), getProperties(value)]
-      })
-    )
-
-    addUtilities(utilities, variants('textColor'))
   }
 }
