+++ src/plugins/transitionDuration.js
@@ -1,7 +1,5 @@
 import createUtilityPlugin from '../util/createUtilityPlugin'
 
 export default function () {
-  return createUtilityPlugin('transitionDuration', [['duration', ['transitionDuration']]], {
-    filterDefault: true,
-  })
+  return createUtilityPlugin('transitionDuration', [['duration', ['transitionDuration']]])
 }
