+++ src/plugins/transitionTimingFunction.js
@@ -1,7 +1,5 @@
 import createUtilityPlugin from '../util/createUtilityPlugin'
 
 export default function () {
-  return createUtilityPlugin('transitionTimingFunction', [['ease', ['transitionTimingFunction']]], {
-    filterDefault: true,
-  })
+  return createUtilityPlugin('transitionTimingFunction', [['ease', ['transitionTimingFunction']]])
 }
